Refactoring some code around.
