+++ typescript/src/lib/jira/extractor.ts
@@ -1,5 +1,5 @@
-import { getAllWorkItemsFromJira } from './repository';
-import { IJiraSettings } from './models'
+import { getAllWorkItemsFromJiraApi } from './api-adapter/main';
+import { IJiraSettings } from './settings'
 import { IWorkItem } from '../core/work-item';
 
 class JiraExtractor {
@@ -13,14 +13,15 @@ class JiraExtractor {
     this.settings = settings;
   }
 
-  async getWorkItems(settings = this.settings): Promise<void> {
-    try {
-      const items = await getAllWorkItemsFromJira(settings);
-      this.workItems = items;
-    } catch (e) {
-      throw e;
-    }
-  }
+  getWorkItems(settings = this.settings): Promise<any> {
+    return new Promise((accept, reject) => {
+      getAllWorkItemsFromJiraApi(settings)
+      .then(items => {
+        this.workItems = items;
+        accept(items);
+      });
+    });
+  };
 
   toCSV(workItems = this.workItems, stages = this.settings.Stages, attributes = this.settings.Attributes) {
     const header = `ID,Link,Name,${stages.join(',')},Type,${Object.keys(attributes).join(',')}`;
@@ -31,5 +32,5 @@ class JiraExtractor {
 };
 
 export {
-  JiraExtractor
+  JiraExtractor,
 };
